requirements.txt
asgiref==3.2.10
Django==3.1.1
django-crispy-forms==1.9.2
djangorestframework==3.12.2
mysql-connector-python==8.0.21
mysqlclient==2.0.1
Pillow==7.2.0
protobuf==3.13.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
